backend/requirements.txt
Flask
flask-cors
requests
beautifulsoup4
pandas
python-dateutil